requirements.txt
streamlit
groq
PyGithub